design : RollingPage 헤더 반응형 기본 구현 완료
