+++ src/components/HeaderService/HeaderService.styled.js
@@ -1,10 +1,24 @@
 import styled from 'styled-components';
+import Inner from '../Inner/Inner';
 
 export const HeaderServiceLayout = styled.div`
   display: flex;
-  padding: 13px 0;
+  height: 68px;
   align-items: center;
   justify-content: space-between;
+
+  @media (max-width: 768px) {
+    flex-direction: column;
+    align-items: start;
+    height: auto;
+    padding: 0;
+  }
+`;
+
+export const InnerContainer = styled(Inner)`
+  @media (max-width: 768px) {
+    width: 100%;
+  }
 `;
 
 export const HeaderServiceTitleBox = styled.h1`
@@ -12,13 +26,17 @@ export const HeaderServiceTitleBox = styled.h1`
   color: var(--gray-800, #2b2b2b);
   font-family: Pretendard;
   font-size: 28px;
-  font-style: normal;
   font-weight: 700;
   line-height: 42px;
   letter-spacing: -0.28px;
 
   @media (max-width: 768px) {
-    display: none;
+    width: 100%;
+    padding: 12px 20px;
+    border-bottom: 1px solid var(--gray-200, #eee);
+    font-size: 18px;
+    line-height: 28px;
+    letter-spacing: -0.18px;
   }
 `;
 
@@ -28,6 +46,8 @@ export const HeaderServiceContainer = styled.div`
   gap: 28px;
 
   @media (max-width: 768px) {
+    width: 100%;
+    padding: 12px 20px;
     gap: 0;
   }
 `;
@@ -69,18 +89,27 @@ export const BarItemsInner = styled.div`
   width: 1px;
   height: 28px;
   background: var(--gray-200, #eee);
+
+  @media (max-width: 1248px) {
+    display: none;
+  }
 `;
 
 export const EmojiContainer = styled.div`
   display: flex;
   align-items: flex-start;
   gap: 8px;
+  @media (max-width: 768px) {
+    width: 100%;
+    justify-content: space-between;
+  }
 `;
 
 export const EmojiBadgeContainer = styled.div`
   display: flex;
   align-items: flex-start;
   gap: 2px;
+  position: relative;
 `;
 
 export const EmojiButtonContainer = styled.div`
