+++ src/components/Inner/Inner.jsx
@@ -1,11 +1,6 @@
 import React from 'react';
 import * as S from './Inner.styled';
 
-export default function Inner({ children }) {
-  return (
-    <S.InnerLayout>
-      { children }
-    </S.InnerLayout>
-  );
+export default function Inner({ children, className }) {
+  return <S.InnerLayout className={className}>{children}</S.InnerLayout>;
 }
-
